# documents/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Nexus API",
    "description": "API платформы Nexus.\n\nКак авторизоваться: выполните `POST /api/v1/auth/login` (тестовые пользователи: `admin`, `author`, `user`, пароль `password123`), скопируйте `access_token` и вставьте его в **Authorize** справа вверху. Токен живёт 15 минут.\n\nПисьма (подтверждение email, вход по ссылке, сброс пароля) в dev-окружении приходят в Mailpit: http://localhost:8025. Токен — параметр `token=` в ссылке из письма.",
    "license": {
      "name": "proprietary",
      "identifier": "proprietary"
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/auth/email/login": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Вход по почте без пароля: шаг 1 — отправить ссылку.",
        "description": "Работает и для новых адресов: аккаунт создаётся, когда пользователь перейдёт по ссылке.",
        "operationId": "email_login_start",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EmailRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Ссылка для входа отправлена"
          },
          "400": {
            "description": "Невалидный email",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "429": {
            "description": "Слишком много писем на этот адрес",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/email/login/confirm": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Вход по почте без пароля: шаг 2 — токен из ссылки. Новый адрес — создаётся аккаунт.",
        "operationId": "email_login_confirm",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EmailTokenRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Вход выполнен (аккаунт создан, если его не было)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenResponse"
                }
              }
            }
          },
          "400": {
            "description": "Токен недействителен, использован или истёк",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "401": {
            "description": "Аккаунт заблокирован",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/email/verify": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Подтверждение email по ссылке из письма после регистрации. Сразу выполняет вход.",
        "operationId": "verify_email",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EmailTokenRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Email подтверждён, пара токенов",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenResponse"
                }
              }
            }
          },
          "400": {
            "description": "Токен недействителен, использован или истёк",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "401": {
            "description": "Аккаунт заблокирован",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/email/verify/resend": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Отправить письмо для подтверждения email ещё раз.",
        "operationId": "resend_verification",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EmailRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Если адрес зарегистрирован и не подтверждён — письмо отправлено"
          },
          "429": {
            "description": "Слишком много писем на этот адрес",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/auth/refresh": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Новая пара токенов по refresh-токену.",
        "description": "Refresh-токен одноразовый: старый отзывается, выдаётся новый. Повторное использование уже\nзаменённого токена значит, что его украли (или клиент сломан), — тогда отзываются все сессии.",
        "operationId": "refresh",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Новая пара токенов",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenResponse"
                }
              }
            }
          },
          "401": {
            "description": "Токен недействителен, истёк или отозван",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "EmailRequest": {
        "type": "object",
        "required": [
          "email"
        ],
        "properties": {
          "email": {
            "type": "string",
            "example": "neo@example.com"
          }
        }
      },
      "EmailTokenRequest": {
        "type": "object",
        "description": "Токен из ссылки в письме (параметр `token=`).",
        "required": [
          "token"
        ],
        "properties": {
          "token": {
            "type": "string"
          }
        }
      },
      "ErrorBody": {
        "type": "object",
        "description": "Тело любого ответа с ошибкой. Нужно для документации API.",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "string",
            "description": "Описание ошибки.",
            "example": "unauthorized"
          }
        }
      },
...
      "RefreshRequest": {
        "type": "object",
        "required": [
          "refresh_token"
        ],
        "properties": {
          "refresh_token": {
            "type": "string"
          }
        }
      },
...
      "Role": {
        "type": "string",
        "description": "Роль пользователя. Порядок вариантов задаёт старшинство: `User < Author < Admin`,\nтак что проверка «не ниже автора» — это `role >= Role::Author`.",
        "enum": [
          "user",
          "author",
          "admin"
        ]
      },
...
      "TokenResponse": {
        "type": "object",
        "description": "Пара токенов. Access — в заголовок `Authorization: Bearer ...`, refresh — для `/refresh`.",
        "required": [
          "access_token",
          "token_type",
          "expires_in",
          "refresh_token",
          "user"
        ],
        "properties": {
          "access_token": {
            "type": "string"
          },
          "expires_in": {
            "type": "integer",
            "format": "int64",
            "description": "Через сколько секунд истекает access-токен.",
            "example": 900
          },
          "refresh_token": {
            "type": "string"
          },
          "token_type": {
            "type": "string",
            "example": "Bearer"
          },
          "user": {
            "$ref": "#/components/schemas/UserView"
          }
        }
      },
      "UserView": {
        "type": "object",
        "description": "Пользователь. Хеш пароля наружу не отдаётся никогда.",
        "required": [
          "id",
          "email",
          "username",
          "role",
          "is_active",
          "created_at"
        ],
        "properties": {
          "avatar_url": {
            "type": [
              "string",
              "null"
            ]
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "display_name": {
            "type": [
              "string",
              "null"
            ]
          },
          "email": {
            "type": "string",
            "example": "user@nexus.local"
          },
          "email_verified_at": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time",
            "description": "`null` — email не подтверждён, вход по паролю запрещён."
          },
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "is_active": {
            "type": "boolean",
            "description": "`false` — заблокирован админом."
          },
          "role": {
            "$ref": "#/components/schemas/Role"
          },
          "username": {
            "type": "string",
            "example": "user"
          }
        }
      }
    },
    "securitySchemes": {
      "bearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
...
}